+++ CHANGELOG.md
@@ -44,6 +44,8 @@
   action will not be sent to the application
   (https://codeberg.org/dnkl/foot/issues/765).
 * Color schemes are now installed to `${datadir}/foot/themes`.
+* `[csd].border-width` and `[csd].border-color`, allowing you to
+  configure the width and color of the CSD border.
 
 
 ### Changed